Restructuring Swift wrapper  (#10)
1. Clean up the sharp edges of the API
a. Expose the user path as a separate config option to allow for consuming apps to store the base dictionaries and user-generated dictionaries in separate locations.
b. Expose the return values of the library functions through the simplified wrapper and the Swift wrapper to allow for fine-grain error handling.
2. Resolve synchronisation issues in the Swift wrapper
a. Add queue-based synchronisation to logger operation to avoid race conditions when the library produces multiple log messages on different threads.
b. Made the Swift Chewing wrapper an actor to leverage Swift's concurrency model for synchronising shared state and callback management.
3. Resolve packaging issues with the `xcframework`
a. Update CMake responsible for generating xcframework so it first packages the compiled dylibs into properly structured frameworks with modulemap files, Info.plist with all information required by Apple and updated install name to match the new structure.
b. Remove precompiled doccarchive from the repo as it will be generated during the release process at a later point.
c. Remove precompiled xcframework from the repo, as it should be produced as part of the build process.
d. Move the Package.swift file to the root of the repository and update the paths accordingly.

---------

Co-authored-by: Yani Yankov <yyankov@abaltatech.com>
Co-authored-by: dkyurtov <dkyurtov@abaltatech.com>